Split auth logic to controller file.
